base class for error exception
